DiffGraphBuilder: make everything string-based (#107)
* cleanup/refactor: use Int everywhere, use `Kind` instead of `Id` in names

* DiffGraphBuilder: make everything string-based

tests don't compile yet, this is an RFC for now

* Cpg.newDiffGraphBuilder

* rename: schema: getPropertyKindByLabel -> getPropertyKindByName

* we need NodeXyz.PropertyNames, not PropertyKinds again now

* fixup

* PR discussions

* whoopsie

* comment re addEdge and property

* valid range is [0, Short.MaxValue]

* throw specific ConversionException, also be more deliberate about lazyness of assert message creation

* fmt

+++ core/src/main/scala/io/joern/odb2/DiffGraphApplier.scala
@@ -1,5 +1,6 @@
 package io.joern.odb2
 
+import DiffGraphBuilder.*
 import io.joern.odb2.Edge.Direction
 import io.joern.odb2.Edge.Direction.{Incoming, Outgoing}
 
@@ -633,9 +634,9 @@ private[odb2] class DiffGraphApplier(graph: Graph, diff: DiffGraphBuilder) {
 }
 
 /* Internal representations of half-edges. These are always normalized to represent either src <- dst or src -> dst. */
-private[odb2] class AddEdgeProcessed(val src: GNode, val dst: GNode, val edgeKind: Short, val property: Any)
+private[odb2] class AddEdgeProcessed(val src: GNode, val dst: GNode, val edgeKind: Int, val property: Any)
 
-private[odb2] class EdgeRepr(val src: GNode, val dst: GNode, val edgeKind: Short, val subSeq: Int, val property: Any)
+private[odb2] class EdgeRepr(val src: GNode, val dst: GNode, val edgeKind: Int, val subSeq: Int, val property: Any)
 
 private[odb2] class SetPropertyDesc(val node: GNode, val start: Int, val end: Int) {
   def length: Int = end - start
